# Vectorize DivideByInPlaceImplementation with Vector<T>