--- a/dotnet/src/SemanticKernel/AI/Embeddings/VectorOperations/DivideOperation.cs
+++ b/dotnet/src/SemanticKernel/AI/Embeddings/VectorOperations/DivideOperation.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using System;
+using System.Numerics;
 using System.Runtime.InteropServices;
 
 namespace Microsoft.SemanticKernel.AI.Embeddings.VectorOperations;
@@ -54,10 +55,24 @@ public static class DivideOperation
         fixed (float* pxBuffer = x)
         {
             float* px = pxBuffer;
-            float* pxMax = px + x.Length;
-            while (px < pxMax)
+            float* pxEnd = px + x.Length;
+
+            if (Vector.IsHardwareAccelerated &&
+                x.Length >= Vector<float>.Count)
             {
-                *px = *px / divisor;
+                Vector<float> divisorVec = new Vector<float>(divisor);
+                float* pxOneVectorFromEnd = pxEnd - Vector<float>.Count;
+                do
+                {
+                    *(Vector<float>*)px /= divisorVec;
+                    px += Vector<float>.Count;
+                }
+                while (px <= pxOneVectorFromEnd);
+            }
+
+            while (px < pxEnd)
+            {
+                *px /= divisor;
                 px++;
             }
         }
@@ -68,10 +83,24 @@ public static class DivideOperation
         fixed (double* pxBuffer = x)
         {
             double* px = pxBuffer;
-            double* pxMax = px + x.Length;
-            while (px < pxMax)
+            double* pxEnd = px + x.Length;
+
+            if (Vector.IsHardwareAccelerated &&
+                x.Length >= Vector<double>.Count)
             {
-                *px = *px / divisor;
+                Vector<double> divisorVec = new Vector<double>(divisor);
+                double* pxOneVectorFromEnd = pxEnd - Vector<double>.Count;
+                do
+                {
+                    *(Vector<double>*)px /= divisorVec;
+                    px += Vector<double>.Count;
+                }
+                while (px <= pxOneVectorFromEnd);
+            }
+
+            while (px < pxEnd)
+            {
+                *px /= divisor;
                 px++;
             }
         }